+++ tests/TestCase/Utility/AbstractBackupUtilityTest.php
@@ -15,9 +15,13 @@ declare(strict_types=1);
 
 namespace DatabaseBackup\Test\TestCase\Utility;
 
+use BadMethodCallException;
 use DatabaseBackup\Driver\AbstractDriver;
 use DatabaseBackup\TestSuite\TestCase;
 use DatabaseBackup\Utility\AbstractBackupUtility;
+use PHPUnit\Framework\Attributes\Test;
+use PHPUnit\Framework\Attributes\TestWith;
+use PHPUnit\Framework\Attributes\WithoutErrorHandler;
 
 /**
  * AbstractBackupUtilityTest.
@@ -26,6 +30,60 @@ use DatabaseBackup\Utility\AbstractBackupUtility;
  */
 class AbstractBackupUtilityTest extends TestCase
 {
+    /**
+     * @uses \DatabaseBackup\Utility\AbstractBackupUtility::__call()
+     */
+    #[Test]
+    public function testMagicCallMethod(): void
+    {
+        $Utility = new class extends AbstractBackupUtility {
+            public function __construct()
+            {
+                $this->filename = 'thisIsAFilename';
+            }
+
+            public function filename(string $filename): AbstractBackupUtility
+            {
+                $this->filename = $filename;
+
+                return $this;
+            }
+        };
+
+        $this->assertSame('thisIsAFilename', $Utility->getFilename());
+    }
+
+    /**
+     * @uses \DatabaseBackup\Utility\AbstractBackupUtility::__call()
+     */
+    #[Test]
+    #[TestWith(['getNoExistingProperty'])]
+    #[TestWith(['noExistingMethod'])]
+    public function testMagicCallMethodWithNoExistingMethod(string $noExistingMethod): void
+    {
+        $this->expectException(BadMethodCallException::class);
+        $this->expectExceptionMessageMatches('/^Method `MockObject_AbstractBackupUtility_\w+::' . $noExistingMethod . '\(\)` does not exist\.$/');
+        $Utility = $this->getMockBuilder(AbstractBackupUtility::class)
+            ->onlyMethods(['filename'])
+            ->getMock();
+        $Utility->{$noExistingMethod}();
+    }
+
+    /**
+     * @uses \DatabaseBackup\Utility\AbstractBackupUtility::__get()
+     */
+    #[Test]
+    #[WithoutErrorHandler]
+    public function testMagicGetMethodIsDeprecated(): void
+    {
+        $Utility = $this->createPartialMock(AbstractBackupUtility::class, ['filename']);
+        $Utility->timeout(3);
+
+        $this->deprecated(function () use ($Utility): void {
+            $Utility->timeout;
+        });
+    }
+
     /**
      * @test
      * @uses \DatabaseBackup\Utility\AbstractBackupUtility::getDriver()